test: 로또 번호 일치 갯수 테스트 추출
